Moved Replay from util library to umpire library since it uses objects that are in the umpire library (Allocator).
