+++ src/umpire/util/CMakeLists.txt
@@ -18,8 +18,7 @@ set (umpire_util_headers
   Exception.hpp
   Logger.hpp
   Macros.hpp
-  Platform.hpp
-  Replay.hpp)
+  Platform.hpp)
 
 if (ENABLE_STATISTICS)
   set (umpire_util_headers
@@ -32,8 +31,7 @@ endif()
 set (umpire_util_sources
   AllocationMap.cpp
   Exception.cpp
-  Logger.cpp
-  Replay.cpp)
+  Logger.cpp)
 
 if (ENABLE_STATISTICS)
   set (umpire_util_sources
