+++ src/umpire/util/Macros.hpp
@@ -119,25 +119,6 @@
 
 #endif // defined(UMPIRE_ENABLE_STATISTICS)
 
-#include "umpire/util/Replay.hpp"
-#define UMPIRE_REPLAY( msg )                                             \
-{                                                                        \
-  if (umpire::util::Replay::getReplayLogger()->replayLoggingEnabled()) { \
-    std::ostringstream local_msg;                                        \
-    local_msg  << "REPLAY," << msg;                                      \
-    umpire::util::Replay::getReplayLogger()->logMessage(local_msg.str());\
-  }                                                                      \
-}
-
-#define UMPIRE_REPLAY_CONT( msg )                                        \
-{                                                                        \
-  if (umpire::util::Replay::getReplayLogger()->replayLoggingEnabled()) { \
-    std::ostringstream local_msg;                                        \
-    local_msg  << "," << msg;                                            \
-    umpire::util::Replay::getReplayLogger()->logMessage(local_msg.str());\
-  }                                                                      \
-}
-
 #define UMPIRE_LOCK \
   if ( !m_mutex->try_lock() ) \
     m_mutex->lock();
